refactor vespa_client into mixins
