+++ src/marqo/vespa/vespa_client/__init__.py
@@ -1,0 +1,1 @@
+from .client import VespaClient
